Remove duplication of console reporters
